Add files via upload
A simple self-made solution by me and my team. All the description are in the pdf inside the file. 
PS: We did this not only for our mini project, but we wanted to make a solution which might be efficient and reliable as government has all the authority over everything.